requirements.txt
# Web framework
Flask==3.0.0

# LINE Messaging API SDK
line-bot-sdk==3.21.0

# Environment variable management
python-dotenv==1.0.0

# HTTP client (for future bus API integration)
requests>=2.31.0
